Split Rule class into SingularRule and CompoundRule types, with the parent Rule type delgating to the concrete types for evaluation and JSON'ing
